flask_postgres/requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.5
bcrypt==4.0.1
